refactor: modularize `handleChat`
